[Modified] OC-959 #comment Merged Invoker sync service with invoker sync scheduler #time 1h
