+++ src/backend/src/main/java/com/becon/opencelium/backend/controller/InvokerController.java
@@ -34,7 +34,6 @@ import com.becon.opencelium.backend.resource.connector.FunctionDTO;
 import com.becon.opencelium.backend.resource.connector.InvokerDTO;
 import com.becon.opencelium.backend.resource.connector.InvokerXMLResource;
 import com.becon.opencelium.backend.resource.error.ErrorResource;
-import com.becon.opencelium.backend.subscription.remoteapi.InvokerSyncScheduler;
 import com.becon.opencelium.backend.utility.PathUtility;
 import com.becon.opencelium.backend.utility.Xml;
 import io.swagger.v3.oas.annotations.Operation;
@@ -81,7 +80,6 @@ import java.util.stream.Collectors;
 public class InvokerController {
     private final InvokerService invokerService;
     private final InvokerSyncService invokerSyncService;
-    private final InvokerSyncScheduler invokerSyncScheduler;
     private final ConnectorService connectorService;
     private final ConnectionService connectionService;
     private final Mapper<Invoker, InvokerDTO> invokerMapper;
@@ -89,15 +87,14 @@ public class InvokerController {
 
     public InvokerController(
             @Qualifier("invokerServiceImp") InvokerService invokerService,
-            InvokerSyncService invokerSyncService, InvokerSyncScheduler invokerSyncScheduler,
             @Qualifier("connectorServiceImp") ConnectorService connectorService,
             @Qualifier("connectionServiceImp") ConnectionService connectionService,
+            InvokerSyncService invokerSyncService,
             Mapper<Invoker, InvokerDTO> invokerMapper,
             Mapper<FunctionInvoker, FunctionDTO> functionMapper
     ) {
         this.invokerService = invokerService;
         this.invokerSyncService = invokerSyncService;
-        this.invokerSyncScheduler = invokerSyncScheduler;
         this.connectorService = connectorService;
         this.connectionService = connectionService;
         this.invokerMapper = invokerMapper;
@@ -409,7 +406,7 @@ public class InvokerController {
     })
     @PutMapping("/{invokerName}/sync-force")
     public ResponseEntity<?> syncForce(@PathVariable String invokerName) {
-        invokerSyncScheduler.forceSync(invokerName);
+        invokerSyncService.forceSync(invokerName);
         return ResponseEntity.noContent().build();
     }
 
@@ -426,7 +423,7 @@ public class InvokerController {
     })
     @PutMapping(path = "list/sync-force", consumes = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<?> syncForceByNames(@RequestBody IdentifiersDTO<String> invokerNames) {
-        invokerNames.getIdentifiers().forEach(invokerSyncScheduler::forceSync);
+        invokerNames.getIdentifiers().forEach(invokerSyncService::forceSync);
         return ResponseEntity.noContent().build();
     }
 
@@ -451,7 +448,7 @@ public class InvokerController {
     }
     
     private InvokerDTO setManualChangeValue(InvokerDTO dto) {
-        dto.setHasManualSync(invokerSyncService.hasManualChange(dto.getName()));
+        dto.setHasManualSync(invokerSyncService.isManuallyModified(dto.getName()));
 
         return dto;
     }
